Significant refactoring
As I look to add more workloads, I realize I needed to redo the
simplistic workload definition so that I can customize what exactly
happens within a workload.

This commit isn't the end game, but it decouples the various pieces such
that it'll be easier to express the counter workload more naturally.